backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.1.0
gunicorn==21.2.0
Werkzeug==3.0.1
requests==2.31.0
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0